Organized the files in different folders + fixed the pet.png
